+++ compression/empty_test.go
