+++ exporter/opencensusexporter/factory.go
@@ -24,13 +24,13 @@ import (
 	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/keepalive"
 
+	"github.com/open-telemetry/opentelemetry-service/compression"
+	compressiongrpc "github.com/open-telemetry/opentelemetry-service/compression/grpc"
 	"github.com/open-telemetry/opentelemetry-service/config/configerror"
 	"github.com/open-telemetry/opentelemetry-service/config/configmodels"
 	"github.com/open-telemetry/opentelemetry-service/consumer"
 	"github.com/open-telemetry/opentelemetry-service/exporter"
 	"github.com/open-telemetry/opentelemetry-service/exporter/exporterhelper"
-	"github.com/open-telemetry/opentelemetry-service/internal/compression"
-	compressiongrpc "github.com/open-telemetry/opentelemetry-service/internal/compression/grpc"
 )
 
 const (
@@ -61,62 +61,15 @@ func (f *Factory) CreateDefaultConfig() configmodels.Exporter {
 // CreateTraceExporter creates a trace exporter based on this config.
 func (f *Factory) CreateTraceExporter(logger *zap.Logger, config configmodels.Exporter) (consumer.TraceConsumer, exporter.StopFunc, error) {
 	ocac := config.(*Config)
+	opts, err := f.OCAgentOptions(logger, ocac)
+	if err != nil {
+		return nil, nil, err
+	}
+	return f.CreateOCAgent(logger, ocac, opts)
+}
 
-	if ocac.Endpoint == "" {
-		return nil, nil, &ocTraceExporterError{
-			code: errEndpointRequired,
-			msg:  "OpenCensus exporter config requires an Endpoint",
-		}
-	}
-
-	opts := []ocagent.ExporterOption{ocagent.WithAddress(ocac.Endpoint)}
-	if ocac.Compression != "" {
-		if compressionKey := compressiongrpc.GetGRPCCompressionKey(ocac.Compression); compressionKey != compression.Unsupported {
-			opts = append(opts, ocagent.UseCompressor(compressionKey))
-		} else {
-			return nil, nil, &ocTraceExporterError{
-				code: errUnsupportedCompressionType,
-				msg:  fmt.Sprintf("OpenCensus exporter unsupported compression type %q", ocac.Compression),
-			}
-		}
-	}
-	if ocac.CertPemFile != "" {
-		creds, err := credentials.NewClientTLSFromFile(ocac.CertPemFile, "")
-		if err != nil {
-			return nil, nil, &ocTraceExporterError{
-				code: errUnableToGetTLSCreds,
-				msg:  fmt.Sprintf("OpenCensus exporter unable to read TLS credentials from pem file %q: %v", ocac.CertPemFile, err),
-			}
-		}
-		opts = append(opts, ocagent.WithTLSCredentials(creds))
-	} else if ocac.UseSecure {
-		certPool, err := x509.SystemCertPool()
-		if err != nil {
-			return nil, nil, &ocTraceExporterError{
-				code: errUnableToGetTLSCreds,
-				msg: fmt.Sprintf(
-					"OpenCensus exporter unable to read certificates from system pool: %v", err),
-			}
-		}
-		creds := credentials.NewClientTLSFromCert(certPool, "")
-		opts = append(opts, ocagent.WithTLSCredentials(creds))
-	} else {
-		opts = append(opts, ocagent.WithInsecure())
-	}
-	if len(ocac.Headers) > 0 {
-		opts = append(opts, ocagent.WithHeaders(ocac.Headers))
-	}
-	if ocac.ReconnectionDelay > 0 {
-		opts = append(opts, ocagent.WithReconnectionPeriod(ocac.ReconnectionDelay))
-	}
-	if ocac.KeepaliveParameters != nil {
-		opts = append(opts, ocagent.WithGRPCDialOption(grpc.WithKeepaliveParams(keepalive.ClientParameters{
-			Time:                ocac.KeepaliveParameters.Time,
-			Timeout:             ocac.KeepaliveParameters.Timeout,
-			PermitWithoutStream: ocac.KeepaliveParameters.PermitWithoutStream,
-		})))
-	}
-
+// CreateOCAgent takes ocagent exporter options and create an OC exporter
+func (f *Factory) CreateOCAgent(logger *zap.Logger, ocac *Config, opts []ocagent.ExporterOption) (consumer.TraceConsumer, exporter.StopFunc, error) {
 	numWorkers := defaultNumWorkers
 	if ocac.NumWorkers > 0 {
 		numWorkers = ocac.NumWorkers
@@ -143,6 +96,64 @@ func (f *Factory) CreateTraceExporter(logger *zap.Logger, config configmodels.Ex
 	}
 
 	return oexp, oce.stop, nil
+}
+
+// OCAgentOptions takes the oc exporter Config and generates ocagent Options
+func (f *Factory) OCAgentOptions(logger *zap.Logger, ocac *Config) ([]ocagent.ExporterOption, error) {
+	if ocac.Endpoint == "" {
+		return nil, &ocTraceExporterError{
+			code: errEndpointRequired,
+			msg:  "OpenCensus exporter config requires an Endpoint",
+		}
+	}
+	opts := []ocagent.ExporterOption{ocagent.WithAddress(ocac.Endpoint)}
+	if ocac.Compression != "" {
+		if compressionKey := compressiongrpc.GetGRPCCompressionKey(ocac.Compression); compressionKey != compression.Unsupported {
+			opts = append(opts, ocagent.UseCompressor(compressionKey))
+		} else {
+			return nil, &ocTraceExporterError{
+				code: errUnsupportedCompressionType,
+				msg:  fmt.Sprintf("OpenCensus exporter unsupported compression type %q", ocac.Compression),
+			}
+		}
+	}
+	if ocac.CertPemFile != "" {
+		creds, err := credentials.NewClientTLSFromFile(ocac.CertPemFile, "")
+		if err != nil {
+			return nil, &ocTraceExporterError{
+				code: errUnableToGetTLSCreds,
+				msg:  fmt.Sprintf("OpenCensus exporter unable to read TLS credentials from pem file %q: %v", ocac.CertPemFile, err),
+			}
+		}
+		opts = append(opts, ocagent.WithTLSCredentials(creds))
+	} else if ocac.UseSecure {
+		certPool, err := x509.SystemCertPool()
+		if err != nil {
+			return nil, &ocTraceExporterError{
+				code: errUnableToGetTLSCreds,
+				msg: fmt.Sprintf(
+					"OpenCensus exporter unable to read certificates from system pool: %v", err),
+			}
+		}
+		creds := credentials.NewClientTLSFromCert(certPool, "")
+		opts = append(opts, ocagent.WithTLSCredentials(creds))
+	} else {
+		opts = append(opts, ocagent.WithInsecure())
+	}
+	if len(ocac.Headers) > 0 {
+		opts = append(opts, ocagent.WithHeaders(ocac.Headers))
+	}
+	if ocac.ReconnectionDelay > 0 {
+		opts = append(opts, ocagent.WithReconnectionPeriod(ocac.ReconnectionDelay))
+	}
+	if ocac.KeepaliveParameters != nil {
+		opts = append(opts, ocagent.WithGRPCDialOption(grpc.WithKeepaliveParams(keepalive.ClientParameters{
+			Time:                ocac.KeepaliveParameters.Time,
+			Timeout:             ocac.KeepaliveParameters.Timeout,
+			PermitWithoutStream: ocac.KeepaliveParameters.PermitWithoutStream,
+		})))
+	}
+	return opts, nil
 }
 
 // CreateMetricsExporter creates a metrics exporter based on this config.
